firestore things on explore page
